fix: update types export
