+++ packages/plugin-request/src/index.ts
@@ -15,7 +15,7 @@ export default async function (api) {
   applyMethod('addIceExport', { source: './request/useRequest', exportName: 'useRequest' });
 
   // add iceTypes exports
-  applyMethod('addIceTypesExport', { source: './request/types', specifier: '{ IRequest }', exportName: 'request?: IRequest' });
+  applyMethod('addIceTypesExport', { source: './request/types/index', specifier: '{ IRequest }', exportName: 'request?: IRequest' });
 
   onGetWebpackConfig((config) => {
     // add alias for module.ts use $ice/axiosInstance
