+++ packages/plugin-store/src/template/types.ts.ejs
@@ -1,4 +1,4 @@
-import * as types from '../appStore';
+import * as types from '.';
 
 export type IRootDispatch = types.IRootDispatch;
 export type IRootState = types.IRootState;
